MDEV-18450 Slaves wait shutdown
The patches features an optional shutdown behavior to hold on
until after all connected slaves have been sent the last binlogged event.

The solution therefore disallows killing the dump thread until is
has found EOF of the latest binlog file.
It is up to the shutdown requester (DBA) to set up a sufficiently large
shutdown timeout value in order to enable even a very lagging behind slave
to synchronize. On other hand if a specific slave should be excluded from
this synchronization the DBA has to stop it manually which would
terminate its dump thread as well.

`mysqladmin shutdown' is extended with a `--wait_for_slaves' option to enable
the feature on the client side.